move call_endpoint to correct file and remove old cache_runner script
