+++ src/holon/cache/cache_runner/call_endpoint.py
@@ -1,15 +1,35 @@
 from holon.cache.cache_runner.config import Config
 from holon.serializers.interactive_element import InteractiveElementInput
+from holon.views import HolonV2Service
+from django.http import HttpRequest
 
 
-def call_holon_endpoint(holon_input_configuration: tuple[InteractiveElementInput]):
+def call_holon_endpoint(
+    scenario_id: int, holon_input_configuration: tuple[InteractiveElementInput]
+):
 
-    request_body = str(
-        [
-            f"{interactive_element_input.interactive_element}: {interactive_element_input.value}"
+    request_body = {
+        "scenario": scenario_id,
+        "interactive_elements": [
+            {
+                "interactive_element": interactive_element_input.interactive_element.id,
+                "value": interactive_element_input.value,
+            }
             for interactive_element_input in holon_input_configuration
-        ]
-    )
+        ],
+    }
     Config.logger.log_print(f"Calling HolonV2Service endpoint with configuration {request_body}")
 
-    # TODO call endpoint
+    # Create request to holon endpoint
+    request = HttpRequest()
+    request.data = request_body
+    request.query_params = {}
+    result = HolonV2Service().post(request)
+
+    if result.status_code == 200:
+        Config.logger.log_print("success")
+    else:
+        print(result, result.__dict__)
+        Config.logger.log_print(
+            f"Calling HolonV2Service endpoint failed with response {request.data}"
+        )
